+++ tests/HTTP/SapiTest.php
@@ -144,6 +144,7 @@ class SapiTest extends \PHPUnit\Framework\TestCase
 
     /**
      * @runInSeparateProcess
+     *
      * @depends testSend
      */
     public function testSendLimitedByContentLengthString(): void
@@ -179,6 +180,7 @@ class SapiTest extends \PHPUnit\Framework\TestCase
 
     /**
      * @runInSeparateProcess
+     *
      * @depends testSend
      */
     public function testSendLimitedByContentLengthStream(): void
@@ -203,7 +205,9 @@ class SapiTest extends \PHPUnit\Framework\TestCase
 
     /**
      * @runInSeparateProcess
+     *
      * @depends testSend
+     *
      * @dataProvider sendContentRangeStreamData
      */
     public function testSendContentRangeStream(
@@ -277,6 +281,7 @@ class SapiTest extends \PHPUnit\Framework\TestCase
 
     /**
      * @runInSeparateProcess
+     *
      * @depends testSend
      */
     public function testSendWorksWithCallbackAsBody(): void
@@ -294,5 +299,36 @@ class SapiTest extends \PHPUnit\Framework\TestCase
         $result = ob_get_clean();
 
         $this->assertEquals('foo', $result);
+    }
+
+    public function testSendConnectionAborted(): void
+    {
+        $baseUrl = getenv('BASEURL');
+        if (!$baseUrl) {
+            $this->markTestSkipped('Set an environment value BASEURL to continue');
+        }
+
+        $url = rtrim($baseUrl, '/').'/connection_aborted.php';
+        $chunk_size = 4 * 1024 * 1024;
+        $fetch_size = 6 * 1024 * 1024;
+
+        $stream = fopen($url, 'r');
+        $size = 0;
+
+        while ($size <= $fetch_size) {
+            $temp = fread($stream, 8192);
+            if (false === $temp) {
+                break;
+            }
+            $size += strlen($temp);
+        }
+
+        fclose($stream);
+
+        sleep(5);
+
+        $bytes_read = file_get_contents(sys_get_temp_dir().'/dummy_stream_read_counter');
+        $this->assertEquals($chunk_size * 2, $bytes_read);
+        $this->assertGreaterThanOrEqual($fetch_size, $bytes_read);
     }
 }